fileCSV

+++ src/Main.java
@@ -4,8 +4,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         CustomManager manager = new CustomManager();
-        System.out.println("Danh sách:");
-        manager.display();
+        manager.display(manager.getMap());
         while (true) {
             System.out.println("================Menu===============");
             System.out.println("1. Thêm khách hàng");
@@ -15,25 +14,25 @@ public class Main {
             System.out.println("5. Sắp xếp");
             System.out.println("6. Sửa thông tin khách hàng");
             System.out.println("0. Thoát");
-            System.out.println("Nhập lựa chọn : ");
+            System.out.print("Nhập lựa chọn : ");
             int choice = sc.nextInt();
             int key;
             sc.nextLine();
             switch (choice) {
                 case 1:
                     do {
-                        System.out.println("Nhập key : ");
+                        System.out.print("Nhập key : ");
                         key = sc.nextInt();
                         sc.nextLine();
-                        if (manager.map.containsKey(key)) {
+                        if (manager.getMap().containsKey(key)) {
                             System.out.println("Key đã tồn tại yêu cầu nhập lại !");
                         }
                     }
-                    while (manager.map.containsKey(key));
+                    while (manager.getMap().containsKey(key));
                     manager.addCus(key, themCus());
                     break;
                 case 2:
-                    manager.display();
+                    manager.display(manager.getMap());
                     break;
                 case 3:
                     System.out.println("====================");
@@ -41,7 +40,7 @@ public class Main {
                     System.out.println("2. Tìm theo tuổi ");
                     System.out.println("3. Tìm theo địa chỉ ");
                     System.out.println("4. Tìm theo ID ");
-                    System.out.println("Nhập lựa chọn : ");
+                    System.out.print("Nhập lựa chọn : ");
                     int find = sc.nextInt();
                     sc.nextLine();
                     switch (find) {
@@ -51,17 +50,21 @@ public class Main {
                             if (manager.findByName(findName) == null) {
                                 System.out.println("Không tìm thấy");
                             } else {
-                                System.out.println(manager.findByName(findName));
+                                manager.display(manager.findByName(findName));
                             }
                             break;
                         case 2:
-                            System.out.print("Nhập tuổi cần tìm : ");
-                            int findAge = sc.nextInt();
+                            System.out.println("Nhập khoảng tuổi cần tìm ");
+                            System.out.print("Nhập đầu : ");
+                            int findAgeFirst = sc.nextInt();
                             sc.nextLine();
-                            if (manager.findByAge(findAge) == null) {
+                            System.out.print("Nhập đuôi : ");
+                            int findAgeLast = sc.nextInt();
+                            sc.nextLine();
+                            if (manager.findByAge(findAgeFirst,findAgeLast) == null) {
                                 System.out.println("Không tìm thấy");
                             } else {
-                                System.out.println(manager.findByAge(findAge));
+                                manager.display(manager.findByAge(findAgeFirst,findAgeLast));
                             }
                             break;
                         case 3:
@@ -70,25 +73,41 @@ public class Main {
                             if (manager.findByAddress(findAddress) == null) {
                                 System.out.println("Không tìm thấy");
                             } else {
-                                System.out.println(manager.findByAddress(findAddress));
+                                manager.display(manager.findByAddress(findAddress));
                             }
                             break;
                         case 4:
-                            System.out.println("Nhập ID cần tìm : ");
+                            System.out.print("Nhập ID cần tìm : ");
                             String findID = sc.nextLine();
                             if (manager.findByID(findID) == null) {
                                 System.out.println("Không tìm thấy");
                             } else {
-                                System.out.println(manager.findByID(findID));
+                                manager.display(manager.findByID(findID));
                             }
                             break;
                     }
                     break;
                 case 4:
-                    System.out.println("Nhập key của khách hàng cần xóa :");
+                    System.out.print("Nhập key của khách hàng cần xóa :");
                     int deleteId = sc.nextInt();
-                    manager.deleteCus(deleteId);
-                    manager.display();
+                    if (manager.findCusKey(deleteId) == -1) {
+                        System.out.println("Không tìm thấy key");
+                    } else {
+                        System.out.println("Bạn có chắc chắn muốn xóa ?");
+                        System.out.println("1. Yes");
+                        System.out.println("2. No");
+                        int check = sc.nextInt();
+                        switch (check){
+                            case 1:
+                                manager.deleteCus(deleteId);
+                                manager.display(manager.getMap());
+                                break;
+                            case 2:
+                                manager.display(manager.getMap());
+                                break;
+                        }
+                    }
+
                     break;
                 case 5:
                     System.out.println("===================");
@@ -100,25 +119,34 @@ public class Main {
                     switch (sort) {
                         case 1:
                             manager.sortByName();
-                            manager.display();
+                            manager.display(manager.getMap());
                             break;
                         case 2:
                             manager.sortByAge();
-                            manager.display();
+                            manager.display(manager.getMap());
                             break;
                         case 3:
                             manager.sortByNameAndAge();
-                            manager.display();
+                            manager.display(manager.getMap());
                             break;
                     }
                     break;
                 case 6:
-                    System.out.println("Nhập key cần sửa thông tin:");
-                    int editKey = sc.nextInt();
+                    int editKey;
+                    do {
+                        System.out.print("Nhập key cần sửa thông tin:");
+                        editKey = sc.nextInt();
+                        sc.nextLine();
+                        if (!manager.getMap().containsKey(editKey)) {
+                            System.out.println("Key không tồn tại yêu cầu nhập lại !");
+                        }
+                    }
+                    while (!manager.getMap().containsKey(editKey));
                     manager.editCus(editKey, themCus());
-                    manager.display();
+                    manager.display(manager.getMap());
                     break;
                 case 0:
+                    System.out.println("Bái bai :))");
                     System.exit(0);
             }
         }
